fix: learningview
한글, 영어, 연산 카테고리별로 학습 화면을 구분하였습니다
